+++ CHANGELOG.md
@@ -5,7 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+<!--
+Types of changes:
+    Added for new features.
+    Changed for changes in existing functionality.
+    Deprecated for soon-to-be removed features.
+    Removed for now removed features.
+    Fixed for any bug fixes.
+    Security in case of vulnerabilities.
+-->
+
 ## [Unreleased]
+
+### Fixed
+* Fix synchronization of async tasks and HTTPX timeouts by using a semaphore with the
+same value as ``max_connections`` (#197) @RafaelWO
 
 ## [0.3.0] - 2024-03-02
 
@@ -41,7 +55,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Upgrade multiple packages manually (#133) @RafaelWO
 * Bump pytest from 8.0.1 to 8.0.2 (#139) @dependabot
 
-### Fixed 
+### Fixed
 * Fix: Code in Docs (#138) @RafaelWO